+++ crates/boundless-cli/src/bin/boundless.rs
@@ -1449,7 +1449,7 @@ mod tests {
         contracts::{hit_points::default_allowance, RequestStatus},
         selector::is_groth16_selector,
     };
-    use boundless_test_utils::{
+    use boundless_test_utils::market::{
         create_test_ctx, deploy_mock_callback, get_mock_callback_count, TestCtx, ECHO_ID, ECHO_PATH,
     };
     use order_stream::{run_from_parts, AppState, ConfigBuilder};
