+++ crates/boundless-market/src/request_builder/mod.rs
@@ -632,7 +632,7 @@ mod tests {
         providers::{DynProvider, Provider},
         rpc::types::TransactionRequest,
     };
-    use boundless_test_utils::{create_test_ctx, ECHO_ELF};
+    use boundless_test_utils::market::{create_test_ctx, ECHO_ELF};
     use tracing_test::traced_test;
     use url::Url;
