+++ crates/broker/src/lib.rs
@@ -1077,7 +1077,7 @@ pub mod test_utils {
     use alloy::network::Ethereum;
     use alloy::providers::{Provider, WalletProvider};
     use anyhow::Result;
-    use boundless_test_utils::{TestCtx, ASSESSOR_GUEST_PATH, SET_BUILDER_PATH};
+    use boundless_test_utils::market::{TestCtx, ASSESSOR_GUEST_PATH, SET_BUILDER_PATH};
     use tempfile::NamedTempFile;
     use url::Url;
