+++ crates/broker/src/market_monitor.rs
@@ -636,7 +636,7 @@ mod tests {
         },
         input::GuestEnv,
     };
-    use boundless_test_utils::{
+    use boundless_test_utils::market::{
         create_test_ctx, deploy_boundless_market, mock_singleton, TestCtx, ASSESSOR_GUEST_ID,
         ASSESSOR_GUEST_PATH, ECHO_ID,
     };
