+++ examples/composition/apps/src/main.rs
@@ -187,7 +187,7 @@ mod tests {
     use boundless_market::{
         contracts::hit_points::default_allowance, storage::StorageProviderType,
     };
-    use boundless_test_utils::{create_test_ctx, TestCtx};
+    use boundless_test_utils::market::{create_test_ctx, TestCtx};
     use broker::test_utils::BrokerBuilder;
     use test_log::test;
     use tokio::task::JoinSet;
