+++ examples/counter-with-callback/apps/src/main.rs
@@ -141,7 +141,7 @@ mod tests {
     };
     use boundless_market::contracts::hit_points::default_allowance;
     use boundless_market::storage::StorageProviderType;
-    use boundless_test_utils::{create_test_ctx, TestCtx, ECHO_ID};
+    use boundless_test_utils::market::{create_test_ctx, TestCtx, ECHO_ID};
     use broker::test_utils::BrokerBuilder;
     use risc0_zkvm::Digest;
     use test_log::test;
