+++ examples/counter/apps/src/main.rs
@@ -158,7 +158,7 @@ mod tests {
     };
     use boundless_market::contracts::hit_points::default_allowance;
     use boundless_market::storage::StorageProviderType;
-    use boundless_test_utils::{create_test_ctx, TestCtx};
+    use boundless_test_utils::market::{create_test_ctx, TestCtx};
     use broker::test_utils::BrokerBuilder;
     use test_log::test;
     use tokio::task::JoinSet;
